Add integration test to Prometheus endpoint
