+++ docs/devlog.md
@@ -1,8 +1,8 @@
 Documentation of development steps, environment, and dependencies  
 
-Contributors: atimeofday;
-Goals: Create skeleton collector with Prometheus endpoint;
-Issues: https://github.com/perfpod/memory-collector/issues/19
+- Contributors: atimeofday
+- Goals: Create skeleton collector with Prometheus endpoint
+- Issues: https://github.com/perfpod/memory-collector/issues/19
 
 Initial environment and tools:
 ```
@@ -48,7 +48,8 @@ curl http://localhost:2112/metrics
 
 Issue 19 objective 3: Expose the `up` metric with value 1
 ```
-# Created, registered, and set an 'up' metric in func main()
+// Created, registered, and set an 'up' metric in func main()
+
 upMetric := prometheus.NewGauge(prometheus.GaugeOpts{
 	Namespace: 	"perfpod",
 	Subsystem: 	"memory_collector",
@@ -73,7 +74,7 @@ perfpod_memory_collector_up_metric 1
 
 Issue 19 objective 5: Move the code into a function (not `main()`)
 ```
-# Moved Up metric into "func recordMetrics()" and added function call in main()
+// Moved Up metric into "func recordMetrics()" and added function call in main()
 
 func main() {
 	recordMetrics()
@@ -82,10 +83,90 @@ func main() {
 	http.ListenAndServe(":2112", nil)
 }
 
-# Repeated manual verification endpoint query
+// Repeated manual verification endpoint query
 ```
 
 Issue 19 objective 6: Add an integration test that verifies the metrics are up, using client_golang's testutil
 - TO DO
 - May require assistance
+
+---
+
+- Issue 19 split into 5/5 done and new Issue 20
+- Issue 19 5/5 PR opened and merged
+
+---
+
+- Contributors: atimeofday
+- Goals: Add integration test to Prometheus endpoint
+- Issues: https://github.com/perfpod/memory-collector/issues/20
+
+Research & references:
+```
+https://go.dev/doc/tutorial/add-a-test
+https://albertmoreno.dev/posts/testing-prometheus-metrics-in-integration-tests-in-golang/
+https://github.com/prometheus/client_golang/blob/main/prometheus/testutil/testutil.go
+https://github.com/prometheus/client_golang/blob/main/prometheus/testutil/testutil_test.go
+```
+
+```
+go get github.com/prometheus/client_golang/prometheus/testutil 
+go get github.com/stretchr/testify/require
+```
+
+Go test format:
+```
+[filename]_test.go
+
+import(
+	[...]
+)
+
+func [TestFunction](t *testing.T) {
+	// Set test values
+	// Perform test
+}
+
+// Perform more tests
+```
+
+1. Created skeleton test based on examples 
+
+```
+func TestMetricsUp(t *testing.T) {
+	require.Eventuallyf(t, func() bool {
+
+  		// Test values
+  		// ??? expected format
+
+		if err := testutil.ScrapeAndCompare(serverURL+"/metrics", strings.NewReader(expected), metricName); err == nil {
+		    return true
+		} else {
+			t.Log(err.Error())
+			return false
+		}
+	}, time.Second, 100*time.Millisecond, "Could not find metric %s with value %d", metricName, expectedMetricValue)
+}
+```
+
+2. Checked the implementation of the testutil ScrapeAndCompare function, and notably, the implementation of its own integration test.
+3. Located and implemented the exact input template required by the function, then implemented generalized code for the template.
+4. Researched goroutines to allow automatically initializing the (currently local) remote server to be tested.
+
+```
+go main()
+time.Sleep(1 * time.Second)
+```
+
+5. Refined logical flow from example code for improved readability.
+
+---
+
+- Issue 20 done
+
+---
+
+- Contributors: 
+- Goals: 
+- Issues:
